// src/app/(hydrogen)/(product-page)/update-profile/user-details.type.ts
export type UserProfileType = {
  userType?: string;
  registrationDate?: string;
  id?: string;
  fullName?: string;
  email?: string;
  phoneNumber?: string;
  address?: string;
  shipping_BillingAddress?: string;
  dateofBirth?: string;
  gender?: string;
  userImage?: string;
  businessCategoryId?: number;
  businessDocument?: string;
  firebaseId?: string;
  createdBy?: string;
  businessCategory?: string
};